+++ subscriptions/api/exceptions.py
@@ -1,0 +1,5 @@
+from rest_framework.exceptions import APIException
+
+
+class BadRequest(APIException):
+    status_code = 400