+++ subscriptions/exceptions.py
@@ -1,3 +1,10 @@
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .models import Subscription
+
+
 class QuotaLimitExceeded(Exception):
     pass
 
@@ -16,6 +23,12 @@ class ProlongationImpossible(Exception):
 
 class SubscriptionError(Exception):
     pass
+
+
+class RecurringSubscriptionsAlreadyExist(SubscriptionError):
+    def __init__(self, message, subscriptions: list[Subscription]):
+        super().__init__(message)
+        self.subscriptions = subscriptions
 
 
 class PaymentError(Exception):